Make DNS configuration optional
If DNS IP is set, DNS name must also be set.

In case DNS IP is set in pot.default.conf and it is not
matching the network, treat it like unset.

This should avoid breakage/annoyance with new setups while
it should be backwards compatible, so it won't break existing
setups.